requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.0
# 云端部署新增依赖
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
bcrypt>=4.0.0
pyyaml>=6.0.0
pyjwt>=2.8.0
